Properly set ObjectiveC flags in iOS toolchain.
It before seemed to take the C flags for ObjectiveC, but newer versions of
CMake do not, causing errors in some situations. (e.g. enabling bitcode)

Removed the explicit C++ flags for the different configurations. Especially
important is removing of -ffastmath, which should only be used when testing
that any speed improvements are worth the potential errors, NEVER be used
by default.